@@ -3,6 +3,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_phoenix/flutter_phoenix.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:aljouf/constants/primary_black.dart';
@@ -38,9 +39,6 @@ void main() async {
       print('yes notification: ${message.notification}');
     }
   });
-  await GetStorage.init();
-  final getStorage = GetStorage();
-  final String lang = getStorage.read('lang') ?? 'ar';
   SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
     statusBarColor: Colors.transparent,
     systemNavigationBarColor: Colors.white,
@@ -49,12 +47,27 @@ void main() async {
     statusBarBrightness: Brightness.light,
     systemNavigationBarIconBrightness: Brightness.dark,
   ));
-  runApp(MyApp(lang: lang));
+  await GetStorage.init();
+  runApp(Phoenix(child: const MyApp()));
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key, required this.lang});
-  final String lang;
+class MyApp extends StatefulWidget {
+  const MyApp({super.key});
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  final getStorage = GetStorage();
+  late String lang;
+
+  @override
+  void initState() {
+    super.initState();
+    lang = getStorage.read('lang') ?? 'ar';
+  }
+
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
