@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:aljouf/constants/colors.dart';
 import 'package:aljouf/product/models/product.dart';
 import 'package:aljouf/widgets/custom_text.dart';
@@ -16,14 +15,14 @@ class ProductDescriptionScreen extends StatelessWidget {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          CustomText(
-            text: 'description'.tr,
-            color: Colors.black,
-            fontWeight: FontWeight.w500,
-          ),
-          const SizedBox(
-            height: 8,
-          ),
+          // CustomText(
+          //   text: 'description'.tr,
+          //   color: Colors.black,
+          //   fontWeight: FontWeight.w500,
+          // ),
+          // const SizedBox(
+          //   height: 8,
+          // ),
           CustomText(
             text: product.description!.split('&nbsp;').join(),
             color: brownishGrey,
